# apps/api/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Q-CMS API",
    "version": "v1",
    "description": "Block-first, API-first headless CMS. See `API.md` for the full contract and `SPEC.md` for the implementation plan.",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "Q-CMS Engineering",
      "url": "https://q-cms.dev"
    },
    "x-spec-version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://{tenant}.q-cms.dev/api/v1",
      "variables": {
        "tenant": {
          "default": "demo",
          "description": "Tenant slug"
        }
      }
    },
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal access token (qcs_ prefix)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "status",
                "code",
                "title"
              ],
              "properties": {
                "id": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "detail": {
                  "type": "string"
                },
                "source": {
                  "type": "object",
                  "properties": {
                    "pointer": {
                      "type": "string"
                    },
                    "parameter": {
                      "type": "string"
                    }
                  }
                },
                "meta": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "isSuperAdmin": {
            "type": "boolean"
          },
          "totpEnabled": {
            "type": "boolean"
          }
        }
      },
      "Entry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "in_review",
              "approved",
              "published",
              "archived"
            ]
          },
          "locale": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "publishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Email + password login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "totp": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": "string",
                    "refreshToken": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errors": [
                    {
                      "status": "401",
                      "code": "unauthorized",
                      "title": "Authentication required"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errors": [
                    {
                      "status": "422",
                      "code": "validation_failed",
                      "title": "Validation failed"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh access token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errors": [
                    {
                      "status": "401",
                      "code": "unauthorized",
                      "title": "Authentication required"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke the current session",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revoked": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/magic-link": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request a single-use magic link",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sent": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Current user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "errors": [
                    {
                      "status": "401",
                      "code": "unauthorized",
                      "title": "Authentication required"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/singletons/{slug}": {
      "parameters": [
        {
          "name": "slug",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "singletons"
        ],
        "summary": "Get a singleton",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "$ref": "#/components/schemas/Entry"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "singletons"
        ],
        "summary": "Upsert a singleton",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "$ref": "#/components/schemas/Entry"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}